Split OKF private lifecycle tests
